Revise install/uninstall script

+++ install.sh
@@ -35,13 +35,19 @@ shell_precheck
 script_mode=""
 symlink_sh="/usr/local/bin/"
 target_dir="/opt/salomon"
+target="${color_yellow}${target_dir}${color_none}"
+yesno="${color_yellow}Y${color_none}/${color_yellow}N${color_none}"
 
 confirm() {
     echo
     echo $em "${color_lightcyan}SaLoMon install/uninstall script${color_none}"
     echo
-    read -p "This will $1 SaLoMon. Do you wish to proceed (Y/N)? " choice
-    if [ "$choice" != "Y" ] && [ "$choice" != "y" ]; then
+    echo $em "This will $script_action SaLoMon. Do you wish to proceed"\
+             "($yesno)? \c"
+    read choice
+    
+    egrep "^yes$|^y$" -i <<< $choice &>/dev/null
+    if [ $? -ne 0 ]; then
         echo
         echo $em "${color_lightred}Canceled${color_none} on user request."
         echo
@@ -71,8 +77,10 @@ fi
 
 if [ "$1" = "--uninstall" ] || [ "$1" = "-u" ]; then
     script_mode="uninstall"
+    script_action="${color_lightred}${script_mode}${color_none}"
 elif [ "$1" = "--install" ] || [ "$1" = "-i" ] || [ -z "$1" ]; then
     script_mode="install"
+    script_action="${color_lightgreen}${script_mode}${color_none}"
 else
     error "Invalid argument" "$1"
 fi
@@ -85,16 +93,28 @@ confirm $script_mode
 echo
 echo $em "${color_lightgreen}Started $script_mode process:${color_none}"
 if [ $script_mode = "uninstall" ]; then
-    rm -f ${symlink_sh}/salomon &>/dev/null
-    echo "  - Removed symbolic link for the main script."
+    cd $(pwd | sed -e "s/\/salomon$//g")
 
-    rm -fR $target_dir &>/dev/null
-    echo $em "  - Removed project directory"\
-             "'${color_yellow}${target_dir}${color_none}'."
+    if [ -f ${symlink_sh}/salomon ]; then
+        rm -f ${symlink_sh}/salomon &>/dev/null
+        echo "  - Removed symbolic link for the main script."
+    else
+        echo "  - Symbolic link for the main script does not exist."
+    fi
+
+    if [ -d $target_dir ]; then
+        rm -fR $target_dir &>/dev/null
+        echo $em "  - Removed project directory '${target}'."
+    else
+        echo $em "  - Project directory '${target}' does not exist."
+    fi
 else
-    mkdir -p $target_dir &>/dev/null
-    echo $em "  - Created target directory"\
-             "'${color_yellow}${target_dir}${color_none}'."
+    if [ -d $target_dir ]; then
+        echo $em "  - Target directory '${target}' already exists."
+    else
+        mkdir -p $target_dir &>/dev/null
+        echo $em "  - Created target directory '${target}'."
+    fi
 
     rsync -av $script_dir/* $target_dir/ &>/dev/null
     echo "  - Copied project data to target directory."
@@ -105,8 +125,12 @@ else
     chmod +x $target_dir/*.sh
     echo "  - Set executable flag for main script files."
 
-    ln -s ${target_dir}/salomon.sh ${symlink_sh}/salomon &>/dev/null
-    echo "  - Created symbolic link for the main script."
+    if [ -f ${symlink_sh}/salomon ]; then
+        echo "  - Symbolic link for the main script already exists."
+    else
+        ln -s ${target_dir}/salomon.sh ${symlink_sh}/salomon &>/dev/null
+        echo "  - Created symbolic link for the main script."
+    fi
 fi
 echo $em "${color_lightgreen}Finished $script_mode process.${color_none}"
 echo
